Added Board and relevant tables. Improved representation of classes.
Tables are: board-monster, board-equipment, and board-monster-effect. Purpose is to show active cards on a given board for a given player. Base classes adjusted to show the relationships (simple JOIN embedded as an attribute called boards to show all boards with a given card).
Some class representation (using class.__repr__) was improved to make the result of a print more user friendly.